[fix] id의 타입이 Long 이므로 findById 의 인자를 Long으로 수정

+++ src/main/java/kr/codesqaud/cafe/domain/article/controller/ArticleController.java
@@ -33,7 +33,7 @@ public class ArticleController {
 	}
 
 	@GetMapping("/questions/{id}")
-	public String viewArticle(@PathVariable("id") Integer id, Model model) {
+	public String viewArticle(@PathVariable("id") Long id, Model model) {
 		model.addAttribute("article", articleRepository.findById(id).orElseThrow(NoSuchElementException::new));
 		return "/post/show";
 	}
